Drop react-map-gl-geocoder
