+++ package.json
@@ -31,7 +31,6 @@
     "react-helmet": "^5.2.1",
     "react-i18next": "^11.2.7",
     "react-map-gl": "^5.2.0",
-    "react-map-gl-geocoder": "^2.0.11",
     "react-timeago": "^4.4.0",
     "react-use": "^13.14.3",
     "react-youtube": "^7.9.0",
